refactor(vue): extract indicator management into useIndicatorManager composable
Move all indicator state (mainActiveIndicators, subPanes, indicatorParams,
activeIndicators) and pane management functions (addSubPane, removeSubPane,
handleIndicatorToggle, etc.) into packages/vue/src/composables/chart/useIndicatorManager.ts.
Also migrate ctrl.indicators and ctrl.subPanes subscriptions into the composable
via setupIndicatorSubscriptions(). Reduces KLineChart.vue <script setup> by ~220 lines.
