+++ packages/vue/src/components/KLineChart.vue
@@ -142,9 +142,6 @@ import { provideFullscreenTeleportTarget } from '../composables/useFullscreenTel
 import {
   createChartController,
   type ChartController,
-  type PaneSpec,
-  type IndicatorInstance,
-  type SubIndicatorType,
   type InteractionSnapshot,
   type DrawingToolId,
   type KLineData,
@@ -153,12 +150,9 @@ import {
   kGapFromKWidth,
   DrawingInteractionController,
 } from '@363045841yyt/klinechart-core/controllers'
-import {
-  getRegisteredIndicatorDefinition,
-  getRegisteredIndicatorDefinitions,
-} from '@363045841yyt/klinechart-core/indicators'
 import type { DrawingObject, DrawingStyle } from '@363045841yyt/klinechart-core/plugin'
 import { useChartTheme } from '../composables/chart/useChartTheme'
+import { useIndicatorManager } from '../composables/chart/useIndicatorManager'
 import { SETTINGS_STORAGE_KEY } from '@363045841yyt/klinechart-core/config'
 import LeftToolbar from './LeftToolbar.vue'
 import TopToolbar, { type SymbolItem } from './TopToolbar.vue'
@@ -327,6 +321,16 @@ const selectedDrawingId = ref<string | null>(null)
 const drawings = ref<DrawingObject[]>([])
 const comparisonColorsMap = ref<Map<string, string>>(new Map())
 const comparisonLoading = ref(false)
+
+const {
+  mainActiveIndicators, subActiveIndicators, activeIndicators,
+  indicatorParams, subPanes,
+  buildPaneLayoutIntent, getDefaultParams, isSubPaneIndicator,
+  addSubPane, removeSubPane, clearAllSubPanes,
+  initIndicatorsFromConfig, switchSubIndicator,
+  handleIndicatorToggle, handleUpdateParams, handleReorderSubIndicators,
+  setupIndicatorSubscriptions,
+} = useIndicatorManager(controller, paneRatios)
 
 // 初始化 kWidth / kGap（与 Chart 引擎 zoom→物理值 转换一致）
 const initZoom = zoomLevel.value
@@ -572,221 +576,7 @@ function onScroll() {
   controller.value?.handleScrollEvent()
 }
 
-// 主图指标显式状态（副图指标从 subPanes 派生）
-const mainActiveIndicators = ref<string[]>([])
 
-// 副图指标列表从 subPanes 自动派生
-const subActiveIndicators = computed(() => {
-  const ids: string[] = []
-  const seen = new Set<string>()
-  for (const pane of subPanes.value) {
-    if (!seen.has(pane.indicatorId)) {
-      seen.add(pane.indicatorId)
-      ids.push(pane.indicatorId)
-    }
-  }
-  return ids
-})
-
-// 最终合并列表（主图 + 副图），保持显示顺序
-const activeIndicators = computed(() => [
-  ...mainActiveIndicators.value,
-  ...subActiveIndicators.value,
-])
-
-// 指标参数配置（MA 的 periods 是数组，需要更宽松的类型）
-const indicatorParams = ref<Record<string, Record<string, unknown>>>({})
-
-// 副图槽位状态
-interface SubPaneSlot {
-  id: string // pane ID: 'RSI_0', 'MACD_0', ...
-  indicatorId: SubIndicatorType
-  params: Record<string, unknown>
-}
-
-// 副图槽位数组（支持多副图）
-const subPanes = ref<SubPaneSlot[]>([])
-
-// 最大副图数量
-const maxSubPanes = 4
-
-function buildPaneLayoutIntent(): PaneSpec[] {
-  const mainRatio = paneRatios.value['main'] ?? 3
-  return subPanes.value.length === 0
-    ? [{ id: 'main', ratio: mainRatio, visible: true, role: 'price' }]
-    : [
-        { id: 'main', ratio: mainRatio, visible: true, role: 'price' },
-        ...subPanes.value.map((pane) => ({
-          id: pane.id,
-          ratio: paneRatios.value[pane.id] ?? 1,
-          visible: true,
-          role: 'indicator' as const,
-        })),
-      ]
-}
-
-// 获取指标默认参数
-function getDefaultParams(
-  indicatorId: SubIndicatorType,
-): Record<string, number | boolean | string> {
-  if (indicatorId === 'VOLUME') return {}
-  const meta = getRegisteredIndicatorDefinition(indicatorId)
-  if (meta?.runtime?.defaultConfig) {
-    return { ...meta.runtime.defaultConfig } as Record<string, number | boolean | string>
-  }
-  return {}
-}
-
-// 副图指标判定（基于 registry category + VOLUME 特例）
-function isSubPaneIndicator(id: string): boolean {
-  if (id === 'VOLUME') return true
-  const def = getRegisteredIndicatorDefinition(id)
-  return !!def && def.category !== 'main'
-}
-
-// 添加副图（使用 Chart API）
-function addSubPane(
-  indicatorId: SubIndicatorType = 'VOLUME',
-  params?: Record<string, number | boolean | string>,
-): boolean {
-  if (subPanes.value.length >= maxSubPanes) {
-    return false
-  }
-
-  const mergedParams = params ?? getDefaultParams(indicatorId)
-
-  const paneId = controller.value?.addIndicator(indicatorId, 'sub', mergedParams)
-  if (!paneId) return false
-  return true
-}
-
-function removeSubPane(paneId: string): void {
-  controller.value?.removeIndicator(paneId)
-}
-
-function clearAllSubPanes(): void {
-  for (const pane of subPanes.value) {
-    controller.value?.removeIndicator(pane.id)
-  }
-}
-
-function initIndicatorsFromConfig(): void {
-  const config = props.semanticConfig
-  const c = controller.value
-  if (!config || !c) return
-
-  const mainIndicators = config.indicators?.main
-  if (mainIndicators) {
-    for (const indicator of mainIndicators) {
-      if (indicator.enabled) {
-        const added = c.addIndicator(
-          indicator.type,
-          'main',
-          indicator.params as Record<string, number | boolean | string>,
-        )
-      }
-    }
-  }
-}
-
-function switchSubIndicator(paneId: string, newIndicatorId: SubIndicatorType): void {
-  const nextParams = getDefaultParams(newIndicatorId)
-  controller.value?.replaceSubPaneIndicator(paneId, newIndicatorId, nextParams)
-}
-
-function handleIndicatorToggle(indicatorId: string, active: boolean) {
-  const c = controller.value
-  if (!c) return
-
-  const def = getRegisteredIndicatorDefinition(indicatorId)
-  const isMain = def && (def.category === 'main' || def.allowMainPane)
-  if (isMain) {
-    const existingIndicator = mainActiveIndicators.value.find((id) => id === indicatorId)
-    if (active && !existingIndicator) {
-      c.addIndicator(indicatorId, 'main', indicatorParams.value[indicatorId])
-    } else if (!active && existingIndicator) {
-      c.removeIndicator(indicatorId.toUpperCase())
-    }
-    return
-  }
-
-  if (isSubPaneIndicator(indicatorId)) {
-    if (active) {
-      const existingPane = subPanes.value.find((p) => p.indicatorId === indicatorId)
-      if (existingPane) return
-      if (subPanes.value.length >= maxSubPanes) return
-
-      const paneId = c.addIndicator(indicatorId, 'sub', indicatorParams.value[indicatorId])
-      if (!paneId && subPanes.value.length > 0) {
-        const lastPane = subPanes.value[subPanes.value.length - 1]
-        switchSubIndicator(lastPane.id, indicatorId as SubIndicatorType)
-      }
-    } else {
-      const panesToRemove = subPanes.value.filter((p) => p.indicatorId === indicatorId)
-      panesToRemove.forEach((pane) => {
-        c.removeIndicator(pane.id)
-      })
-    }
-  }
-}
-
-function handleUpdateParams(indicatorId: string, params: Record<string, unknown>) {
-  if (
-    indicatorId === 'MA' ||
-    indicatorId === 'BOLL' ||
-    indicatorId === 'EXPMA' ||
-    indicatorId === 'ENE'
-  ) {
-    controller.value?.updateIndicatorParams(indicatorId, params)
-    return
-  }
-  if (isSubPaneIndicator(indicatorId)) {
-    subPanes.value
-      .filter((p) => p.indicatorId === indicatorId)
-      .forEach((pane) => {
-        controller.value?.updateIndicatorParams(pane.id, params)
-      })
-  }
-}
-
-function handleReorderSubIndicators(orderedIndicatorIds: string[]) {
-  if (!orderedIndicatorIds.length || subPanes.value.length <= 1) return
-
-  const validOrder = orderedIndicatorIds.filter((id): id is SubIndicatorType =>
-    isSubPaneIndicator(id),
-  )
-  if (!validOrder.length) return
-
-  const paneByIndicator = new Map(subPanes.value.map((pane) => [pane.indicatorId, pane] as const))
-  const nextSubPanes: SubPaneSlot[] = []
-
-  for (const indicatorId of validOrder) {
-    const pane = paneByIndicator.get(indicatorId)
-    if (pane) {
-      nextSubPanes.push(pane)
-      paneByIndicator.delete(indicatorId)
-    }
-  }
-
-  if (nextSubPanes.length === 0) return
-
-  for (const pane of subPanes.value) {
-    if (paneByIndicator.has(pane.indicatorId)) {
-      nextSubPanes.push(pane)
-      paneByIndicator.delete(pane.indicatorId)
-    }
-  }
-
-  const currentSubIds = subPanes.value.map((p) => p.id)
-  const nextSubIds = nextSubPanes.map((p) => p.id)
-  if (currentSubIds.join('|') === nextSubIds.join('|')) return
-
-  subPanes.value = nextSubPanes
-
-  const c = controller.value
-  if (!c) return
-  c.updatePaneLayout(buildPaneLayoutIntent())
-}
 
 /* 计算总宽度：从 Vue 响应式状态读取，zoom 变化时自动重算 */
 const axisHostWidth = computed(() => props.rightAxisWidth + props.priceLabelWidth)
@@ -907,58 +697,7 @@ function setupChartCallbacks(ctrl: ChartController): void {
     emit('themeChange', newTheme)
   })
 
-  const unsubscribeIndicators = ctrl.indicators.subscribe(() => {
-    const instances = ctrl.indicators.peek()
-
-    const mains = instances
-      .filter((i): i is IndicatorInstance & { role: 'main' } => i.role === 'main')
-      .map((i) => i.definitionId)
-    mainActiveIndicators.value = mains
-
-    const nextParams = { ...indicatorParams.value }
-    for (const inst of instances) {
-      if (inst.role === 'main' && inst.params && Object.keys(inst.params).length > 0) {
-        nextParams[inst.definitionId] = { ...inst.params }
-      }
-    }
-
-    ctrl.updateRendererConfig('mainIndicatorLegend', {
-      indicators: {
-        MA: { enabled: mains.includes('MA'), params: nextParams['MA'] || {} },
-        BOLL: { enabled: mains.includes('BOLL'), params: nextParams['BOLL'] || {} },
-        EXPMA: { enabled: mains.includes('EXPMA'), params: nextParams['EXPMA'] || {} },
-        ENE: { enabled: mains.includes('ENE'), params: nextParams['ENE'] || {} },
-      },
-    })
-
-    indicatorParams.value = nextParams
-  })
-
-  const unsubscribeSubPanes = ctrl.subPanes.subscribe(() => {
-    const subPaneInfos = ctrl.subPanes.peek()
-    const signalIds = new Set(subPaneInfos.map((sp) => sp.paneId))
-
-    const merged = subPanes.value.filter((p) => signalIds.has(p.id))
-    const existingIds = new Set(merged.map((p) => p.id))
-    for (const sp of subPaneInfos) {
-      if (!existingIds.has(sp.paneId)) {
-        merged.push({
-          id: sp.paneId,
-          indicatorId: sp.indicatorId as SubIndicatorType,
-          params: sp.params,
-        })
-      }
-    }
-    subPanes.value = merged
-
-    const nextParams = { ...indicatorParams.value }
-    for (const sp of subPaneInfos) {
-      if (sp.params && Object.keys(sp.params).length > 0) {
-        nextParams[sp.indicatorId] = { ...sp.params }
-      }
-    }
-    indicatorParams.value = nextParams
-  })
+  const unsubscribeIndicators = setupIndicatorSubscriptions(ctrl)
 
   const unsubscribeComparisonColors = ctrl.comparisonColors.subscribe(() => {
     comparisonColorsMap.value = new Map(ctrl.comparisonColors.peek())
@@ -976,7 +715,6 @@ function setupChartCallbacks(ctrl: ChartController): void {
     unsubscribePaneLayout()
     unsubscribeTheme()
     unsubscribeIndicators()
-    unsubscribeSubPanes()
     unsubscribeComparisonColors()
     unsubscribeComparisonLoading()
   })
@@ -1023,7 +761,7 @@ function setupSemanticController(ctrl: ChartController): void {
 
   // config:ready → Chart 侧已完成创建，Vue 回读状态
   semanticController.value.on('config:ready', () => {
-    initIndicatorsFromConfig()
+    initIndicatorsFromConfig(props.semanticConfig)
     nextTick(() => controller.value?.scrollToRight())
   })
   // 暂时断开语义化配置加载，由搜索结果驱动
@@ -1059,7 +797,7 @@ onMounted(async () => {
   // 3.5) 在任何 draw 之前注册主图指标（BOLL/MA 等）
   //      initIndicatorsFromConfig 是同步的，读 props.semanticConfig 即可注册，
   //      确保 scheduler 首次 applyResults 时 BOLL 已在 registry 里
-  initIndicatorsFromConfig()
+  initIndicatorsFromConfig(props.semanticConfig)
 
   // 4) 工具栏初始设置
   applyInitialSettings(ctrl)
